requirements.txt
boto3==1.37.9
botocore==1.37.9
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.12
cryptography==44.0.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
moto==5.1.1
packaging==24.2
pluggy==1.5.0
pycparser==2.22
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
responses==0.25.6
s3transfer==0.11.4
six==1.17.0
urllib3==2.3.0
Werkzeug==3.1.3
xmltodict==0.14.2
